+++ packages/reactivity-core/src/types.ts
@@ -99,6 +99,18 @@ export interface ExternalReactive<T> extends ReadonlyReactive<T> {
 }
 
 /**
+ * A function that returns a value that may change over time.
+ * 
+ * The function should be implemented in terms of signals (directly or indirectly).
+ * 
+ * Functions of this kind are designed to run in a reactive context.
+ * They will typically be invoked repeatedly if any of their dependencies change.
+ * 
+ * @group Primitives
+ */
+export type ReactiveGetter<T> = () => T;
+
+/**
  * A function that shall return `true` if `a` and `b` are considered equal, `false` otherwise.
  *
  * @group Primitives