+++ packages/reactivity-core/CHANGELOG.md
@@ -2,11 +2,14 @@
 
 ## v0.5.0 (TBD)
 
-- Remove deprecated `syncEffectOnce`
+- Introduce `ReactiveGetter<T>` to mark functions used to compute reactive values.
+  This is used in the signatures of `computed` and `*watch*`.
+  `ReactiveGetter<T>` is a simply type alias for `() => T`.
 - Use `Object.is` instead of `===` for cleaner comparisons (e.g. around NaN)
     - Note that this does not resolve all issues around NaN: a computed signal returning `NaN` will still be considered a change.
       This is due to the underlying implementation in @preact/signals-core, which compares values using `===`.
 - Relax type signature of `watch` and `syncWatch`: the array of values is no longer `readonly` (but it still must not be modified at runtime).
+- Remove deprecated `syncEffectOnce`
 
 ## v0.4.4